requirements.txt
pysistence
bs4
aiohttp
aiofile
python-magic
